requirements.txt
langchain~=0.0.252
python-dotenv>=1.0.0
chromadb~=0.3.21
tiktoken
pandoc
pypandoc~=1.11
dateparser~=1.1.8
streamlit~=1.25.0
langsmith~=0.0.14
EbookLib~=0.18
bs4~=0.0.1
beautifulsoup4~=4.12.2
python-dateutil~=2.8.2
plac~=1.3.5
pandas~=1.5.3
torch~=2.0.0
spacy~=3.5.3
streamlit-chat>=0.0.2.2
streamlit_option_menu>=0.3.2
streamlit_pydantic